backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.35.0
boto3==1.38.34
pydantic==2.11.5
python-dotenv==1.1.1
mangum==0.19.0
python-jose[cryptography]==3.4.0
httpx==0.28.1
